requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
Pillow==7.1.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
tifffile==2020.5.30
torch==1.0.1.post2
torchvision==0.2.2
tqdm==4.46.0
